requirements.txt
alembic==1.0.10
asn1crypto==0.24.0
attrs==19.1.0
blinker==1.4
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.6.1
cycler==0.10.0
decorator==4.4.0
dotenv==0.10.1
Flask==1.0.2
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.8
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
jupyter-core==4.4.0
kiwisolver==1.1.0
Mako==1.0.9
MarkupSafe==1.1.1
matplotlib==3.0.3
nbformat==4.4.0
numpy==1.16.3
pandas==0.24.2
plotly==3.8.1
psycopg2==2.8.2
psycopg2-binary==2.8.2
pycparser==2.19
pydotplus==2.0.2
PyJWT==1.7.1
pyparsing==2.4.0
pyrsistent==0.15.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
requests==2.21.0
retrying==1.3.3
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
SQLAlchemy==1.3.3
traitlets==4.3.2
urllib3==1.21.1
Werkzeug==0.15.2
WTForms==2.2.1
